requirements.txt
certifi==2020.12.5
chardet==3.0.4
config==0.5.0.post0
idna==2.10
paho-mqtt==1.5.1
pymongo==3.11.2
requests==2.25.0
requests-toolbelt==0.9.1
urllib3==1.26.2
